+++ R/register_all_transpilers.R
@@ -16,4 +16,5 @@ register_all_transpilers <- function() {
   append_builtin_transpilers_for_foreach()
   append_builtin_transpilers_for_doFuture()
   append_builtin_transpilers_for_lme4()
+  append_builtin_transpilers_for_SimDesign()
 } ## register_all_transpilers()